recognition/siamese-45033746/requirements.txt
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.1
Pillow==10.1.0
torch==2.1.0
